+++ NEWS.md
@@ -6,6 +6,7 @@
 * added uncoveredMajorityClass.
 * added transactions2DF to convert transactions to a data.frame.
 * RCAR is now faster (does not run glmnet again for the chosen lambda) and returns the whole regularization path.
+* prepareTransactions does now automatically add a negative class item if needed.
 
 # arulesCBA 1.2.0 (2020-04-19)
 * Version 1.2.0 has a major interface cleanup. This might require some change in existing code.